feat: allow to select one of the options as "primary"
Using a double click it is now possible to select one of the options as
primary - its id will be written in a hidden input field and it will be
highlighted with red border.
In addition, the merge button is disabled by default and is only
activated, once a primary is selected.